feat: 更新及删除索引模板接口

+++ component/src/main/java/com/rany/service/component/metric/TemplateMetricCounter.java
@@ -11,7 +11,7 @@ package com.rany.service.component.metric;
  */
 public class TemplateMetricCounter {
 
-    public long totalIndex;
-    public long totalDocs;
-    public long totalDataSize;
+    public Integer totalIndex;
+    public Integer totalDocs;
+    public Integer totalDataSize;
 }
